front: Extract the experience marker builder.
The marker set is a pure function of the experiences, their locations
and which category groups are expanded, but it lived inline in a
useMemo inside a 900-line component, where it could only be exercised
by mounting a map.

Moved to a module of its own, along with the MarkerData shape it
produces. No behaviour change.
